improve toolbar styles
