+++ src/components/Card/Toolbar.tsx
@@ -5,10 +5,3 @@ export const Toolbar: React.FC<PropsWithBox<{ children: React.ReactNode }>> = ({
   children,
   ...rest
 }) => <Box {...rest}>{children}</Box>;
-
-export const toolbarWrapperStyles = {
-  display: "flex",
-  flexDirection: "row",
-  alignItems: "center",
-  justifyContent: "space-between",
-} as const;
