+++ src/discounts/components/DiscountCategories/DiscountCategories.tsx
@@ -2,7 +2,6 @@
 import { categoryUrl } from "@dashboard/categories/urls";
 import { Button } from "@dashboard/components/Button";
 import { DashboardCard } from "@dashboard/components/Card";
-import { toolbarWrapperStyles } from "@dashboard/components/Card/Toolbar";
 import Checkbox from "@dashboard/components/Checkbox";
 import ResponsiveTable from "@dashboard/components/ResponsiveTable";
 import Skeleton from "@dashboard/components/Skeleton";
@@ -46,7 +45,7 @@ const DiscountCategories: React.FC<DiscountCategoriesProps> = props => {
 
   return (
     <DashboardCard data-test-id="assign-category-section">
-      <DashboardCard.Header {...toolbarWrapperStyles}>
+      <DashboardCard.Header withToolbar>
         <DashboardCard.Title>
           {intl.formatMessage(messages.discountCategoriesHeader)}
         </DashboardCard.Title>
