+++ src/discounts/components/DiscountCollections/DiscountCollections.tsx
@@ -2,7 +2,6 @@
 import { collectionUrl } from "@dashboard/collections/urls";
 import { Button } from "@dashboard/components/Button";
 import { DashboardCard } from "@dashboard/components/Card";
-import { toolbarWrapperStyles } from "@dashboard/components/Card/Toolbar";
 import Checkbox from "@dashboard/components/Checkbox";
 import ResponsiveTable from "@dashboard/components/ResponsiveTable";
 import Skeleton from "@dashboard/components/Skeleton";
@@ -46,7 +45,7 @@ const DiscountCollections: React.FC<DiscountCollectionsProps> = props => {
 
   return (
     <DashboardCard data-test-id="assign-collection-section">
-      <DashboardCard.Header {...toolbarWrapperStyles}>
+      <DashboardCard.Header withToolbar>
         <DashboardCard.Title>
           {intl.formatMessage(messages.discountCollectionsHeader)}
         </DashboardCard.Title>
