+++ src/discounts/components/DiscountProducts/DiscountProducts.tsx
@@ -1,7 +1,6 @@
 // @ts-strict-ignore
 import { Button } from "@dashboard/components/Button";
 import { DashboardCard } from "@dashboard/components/Card";
-import { toolbarWrapperStyles } from "@dashboard/components/Card/Toolbar";
 import { ChannelsAvailabilityDropdown } from "@dashboard/components/ChannelsAvailabilityDropdown";
 import Checkbox from "@dashboard/components/Checkbox";
 import ResponsiveTable from "@dashboard/components/ResponsiveTable";
@@ -49,7 +48,7 @@ const DiscountProducts: React.FC<SaleProductsProps> = props => {
 
   return (
     <DashboardCard data-test-id="assign-product-section">
-      <DashboardCard.Header {...toolbarWrapperStyles}>
+      <DashboardCard.Header withToolbar>
         <DashboardCard.Title>
           {intl.formatMessage(messages.discountProductsHeader)}
         </DashboardCard.Title>
