+++ src/permissionGroups/components/PermissionGroupMemberList/PermissionGroupMemberList.tsx
@@ -1,7 +1,6 @@
 // @ts-strict-ignore
 import { Button } from "@dashboard/components/Button";
 import { DashboardCard } from "@dashboard/components/Card";
-import { toolbarWrapperStyles } from "@dashboard/components/Card/Toolbar";
 import Checkbox from "@dashboard/components/Checkbox";
 import ResponsiveTable from "@dashboard/components/ResponsiveTable";
 import Skeleton from "@dashboard/components/Skeleton";
@@ -91,7 +90,7 @@ const PermissionGroupMemberList: React.FC<PermissionGroupProps> = props => {
 
   return (
     <DashboardCard data-test-id="permission-group-members-section">
-      <DashboardCard.Header {...toolbarWrapperStyles}>
+      <DashboardCard.Header withToolbar>
         <DashboardCard.Title>
           {intl.formatMessage({
             id: "lGlDEH",
